Switch from Vue to Svelte for controls UI components
