+++ package.json
@@ -21,22 +21,25 @@
   "dependencies": {
     "comlink": "^4.3.1",
     "pixi.js": "^6.2.0",
-    "vue": "^3.2.23"
+    "svelte": "^3.44.2"
   },
   "devDependencies": {
     "@swc/core": "^1.2.108",
+    "autoprefixer": "^10.4.0",
     "css-loader": "^6.5.1",
     "glslify-loader": "^2.0.0",
     "html-webpack-plugin": "^5.5.0",
+    "mini-css-extract-plugin": "^2.4.5",
     "raw-loader": "^1.0.0",
     "regenerator-runtime": "^0.13.9",
     "resource-loader": "^4.0.0-rc4",
-    "sass": "^1.43.4",
+    "sass": "^1.44.0",
     "sass-loader": "^12.3.0",
     "style-loader": "^3.3.1",
+    "svelte-loader": "^3.1.2",
+    "svelte-preprocess": "^4.9.8",
     "swc-loader": "^0.1.15",
     "typescript": "^4.4.4",
-    "vue-loader": "^16.8.3",
     "webpack": "^5.64.0",
     "webpack-cli": "^4.9.1",
     "webpack-dev-server": "^4.5.0"
