+++ src/index.html
@@ -6,9 +6,6 @@
 </head>
 
 <body>
-  <div id="pixi-root"></div>
-  <div id="ui-root">
-  </div>
 </body>
 
 </html>
