+++ src/index.ts
@@ -1,5 +1,3 @@
-import { createApp } from "vue";
-
 // Import CSS files for webpack build
 import "./style.scss";
 
@@ -7,12 +5,12 @@ import startUpdateLoop from "./update";
 import startInputManager from "./inputManager";
 import startCompletionChecker from "./completionChecker";
 
-import Controls from "./components/Controls.vue";
+import Controls from "./components/Controls.svelte";
 
 startUpdateLoop();
 startInputManager();
 startCompletionChecker();
 
-const controlsVueApp = createApp(Controls);
-
-controlsVueApp.mount("#ui-root");
+new Controls({
+  target: document.body,
+});
