+++ src/style.scss
@@ -6,15 +6,13 @@ body {
   overflow: hidden;
 }
 
-#pixi-root {
+canvas {
+  width: 100%;
+  height: 100%;
+
   cursor: none;
 
-  html[data-ismouseactive="true"] & {
+  html[data-controlsactive="true"] & {
     cursor: auto;
-  }
-
-  canvas {
-    width: 100%;
-    height: 100%;
   }
 }
